% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGADownload.R
\name{TCGA}
\alias{TCGA}
\title{TCGA Download}
\usage{
TCGA()
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
TCGA()
}
\keyword{cats}
