% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{testF}
\alias{testF}
\title{This function allows you to express your love of cats.}
\usage{
testF()
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
testF()
}
\keyword{cats}
